Remove lookup-refs (#989)
Part of https://github.com/insightsengineering/coredev-tasks/issues/609

From now on, we will provide development dependencies in 

```
Remotes: repo/project@branch
```
format, so it's explicitly visible in the DESCRIPTION file and can be
handled by `pak::install`, `renv::install` and `remotes::install`.

With development dependencies specified in CI Pipelines configuration,
this connection was hidden, and it was hard to install the package from
the main branch (or any other branch) locally from user's machine.

formatters was the only package on lookup-refs. It is a dependency
required at the version available on CRAN or higher.