enforce name uniqueness and pathability. unify section_div approach (#1024)
This... ended up being more sizable than I expected to get all the
behavior fixed and unified. Should be nearly backwards compatible with
the exception of behavior that was buggy or otherwise incorrect before.

We should probably meet so I can walk you through this, but I'm happy to
discuss on here instead/in addition.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Joe Zhu <sha.joe.zhu@gmail.com>